Separated out fontutil
